requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
packaging==24.1
pathspec==0.10.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.1.9
